Data in matrix/requirements.txt
pillow
numpy
opencv-python
keras
matplotlib
tensorflow
qrcode